Revert "Adding merge functionality"
