+++ .github/workflows/main.yml
@@ -25,10 +25,10 @@ jobs:
           TWILIO_API_AUTH_TOKEN: ${{ secrets.TWILIO_API_AUTH_TOKEN }}
           TWILIO_API_KEY_SECRET: ${{ secrets.TWILIO_API_KEY_SECRET }}
           TWILIO_API_KEY_SID: ${{ secrets.TWILIO_API_KEY_SID }}
-        run: cd services/roomService; npm install && npm run lint && npm test
+        run: cd services/roomService; npm install && npm test
 
       - name: Build and test frontend components
-        run: cd frontend; npm install && npm run lint && npm test
+        run: cd frontend; npm install && npm test
   deploy:
     if: github.ref == 'refs/heads/master'
     needs: build-and-test
