+++ frontend/src/App.tsx
@@ -3,12 +3,13 @@ import React, {
 } from 'react';
 import './App.css';
 import { BrowserRouter } from 'react-router-dom';
-import { io } from 'socket.io-client';
-import { ChakraProvider, useToast } from '@chakra-ui/react';
+import { io, Socket } from 'socket.io-client';
+import { ChakraProvider } from '@chakra-ui/react';
 import { MuiThemeProvider } from '@material-ui/core/styles';
 import assert from 'assert';
 import WorldMap from './components/world/WorldMap';
 import VideoOverlay from './components/VideoCall/VideoOverlay/VideoOverlay';
+import { CoveyAppState, NearbyPlayers } from './CoveyTypes';
 import VideoContext from './contexts/VideoContext';
 import Login from './components/Login/Login';
 import CoveyAppContext from './contexts/CoveyAppContext';
@@ -22,13 +23,135 @@ import ErrorDialog from './components/VideoCall/VideoFrontend/components/ErrorDi
 import theme from './components/VideoCall/VideoFrontend/theme';
 import { Callback } from './components/VideoCall/VideoFrontend/types';
 import Player, { ServerPlayer, UserLocation } from './classes/Player';
-import { TownJoinResponse } from './classes/TownsServiceClient';
+import TownsServiceClient, { TownJoinResponse } from './classes/TownsServiceClient';
 import Video from './classes/Video/Video';
-import { CoveyAppUpdate, appStateReducer, defaultAppState } from './AppHelper'
+
+type CoveyAppUpdate =
+  | { action: 'doConnect'; data: { userName: string, townFriendlyName: string, townID: string,townIsPubliclyListed:boolean, townIsMergeable:boolean, sessionToken: string, myPlayerID: string, socket: Socket, players: Player[], emitMovement: (location: UserLocation) => void } }
+  | { action: 'addPlayer'; player: Player }
+  | { action: 'playerMoved'; player: Player }
+  | { action: 'playerDisconnect'; player: Player }
+  | { action: 'weMoved'; location: UserLocation }
+  | { action: 'disconnect' }
+  ;
+
+function defaultAppState(): CoveyAppState {
+  return {
+    nearbyPlayers: { nearbyPlayers: [] },
+    players: [],
+    myPlayerID: '',
+    currentTownFriendlyName: '',
+    currentTownID: '',
+    currentTownIsPubliclyListed: false,
+    currentTownIsMergeable: false,
+    sessionToken: '',
+    userName: '',
+    socket: null,
+    currentLocation: {
+      x: 0, y: 0, rotation: 'front', moving: false,
+    },
+    emitMovement: () => {
+    },
+    apiClient: new TownsServiceClient(),
+  };
+}
+function appStateReducer(state: CoveyAppState, update: CoveyAppUpdate): CoveyAppState {
+  const nextState = {
+    sessionToken: state.sessionToken,
+    currentTownFriendlyName: state.currentTownFriendlyName,
+    currentTownID: state.currentTownID,
+    currentTownIsPubliclyListed: state.currentTownIsPubliclyListed,
+    currentTownIsMergeable: state.currentTownIsMergeable,
+    myPlayerID: state.myPlayerID,
+    players: state.players,
+    currentLocation: state.currentLocation,
+    nearbyPlayers: state.nearbyPlayers,
+    userName: state.userName,
+    socket: state.socket,
+    emitMovement: state.emitMovement,
+    apiClient: state.apiClient,
+  };
+
+  function calculateNearbyPlayers(players: Player[], currentLocation: UserLocation) {
+    const isWithinCallRadius = (p: Player, location: UserLocation) => {
+      if (p.location && location) {
+        const dx = p.location.x - location.x;
+        const dy = p.location.y - location.y;
+        const d = Math.sqrt(dx * dx + dy * dy);
+        return d < 80;
+      }
+      return false;
+    };
+    return { nearbyPlayers: players.filter((p) => isWithinCallRadius(p, currentLocation)) };
+  }
+
+  function samePlayers(a1: NearbyPlayers, a2: NearbyPlayers) {
+    if (a1.nearbyPlayers.length !== a2.nearbyPlayers.length) return false;
+    const ids1 = a1.nearbyPlayers.map((p) => p.id).sort();
+    const ids2 = a2.nearbyPlayers.map((p) => p.id).sort();
+    return !ids1.some((val, idx) => val !== ids2[idx]);
+  }
+
+  let updatePlayer;
+  switch (update.action) {
+    case 'doConnect':
+      nextState.sessionToken = update.data.sessionToken;
+      nextState.myPlayerID = update.data.myPlayerID;
+      nextState.currentTownFriendlyName = update.data.townFriendlyName;
+      nextState.currentTownID = update.data.townID;
+      nextState.currentTownIsPubliclyListed = update.data.townIsPubliclyListed;
+      nextState.currentTownIsMergeable = update.data.townIsMergeable;
+      nextState.userName = update.data.userName;
+      nextState.emitMovement = update.data.emitMovement;
+      nextState.socket = update.data.socket;
+      nextState.players = update.data.players;
+      break;
+    case 'addPlayer':
+      nextState.players = nextState.players.concat([update.player]);
+      break;
+    case 'playerMoved':
+      updatePlayer = nextState.players.find((p) => p.id === update.player.id);
+      if (updatePlayer) {
+        updatePlayer.location = update.player.location;
+      } else {
+        nextState.players = nextState.players.concat([update.player]);
+      }
+      nextState.nearbyPlayers = calculateNearbyPlayers(nextState.players,
+        nextState.currentLocation);
+      if (samePlayers(nextState.nearbyPlayers, state.nearbyPlayers)) {
+        nextState.nearbyPlayers = state.nearbyPlayers;
+      }
+      break;
+    case 'weMoved':
+      nextState.currentLocation = update.location;
+      nextState.nearbyPlayers = calculateNearbyPlayers(nextState.players,
+        nextState.currentLocation);
+      if (samePlayers(nextState.nearbyPlayers, state.nearbyPlayers)) {
+        nextState.nearbyPlayers = state.nearbyPlayers;
+      }
+
+      break;
+    case 'playerDisconnect':
+      nextState.players = nextState.players.filter((player) => player.id !== update.player.id);
+
+      nextState.nearbyPlayers = calculateNearbyPlayers(nextState.players,
+        nextState.currentLocation);
+      if (samePlayers(nextState.nearbyPlayers, state.nearbyPlayers)) {
+        nextState.nearbyPlayers = state.nearbyPlayers;
+      }
+      break;
+    case 'disconnect':
+      state.socket?.disconnect();
+      return defaultAppState();
+    default:
+      throw new Error('Unexpected state request');
+  }
+
+  return nextState;
+}
 
 async function GameController(initData: TownJoinResponse,
-  // eslint-disable-next-line
-  dispatchAppUpdate: (update: CoveyAppUpdate) => void, toast?: any) {
+  dispatchAppUpdate: (update: CoveyAppUpdate) => void) {
   // Now, set up the game sockets
   const gamePlayerID = initData.coveyUserID;
   const sessionToken = initData.coveySessionToken;
@@ -57,33 +180,6 @@ async function GameController(initData: TownJoinResponse,
   socket.on('disconnect', () => {
     dispatchAppUpdate({ action: 'disconnect' });
   });
-
-  socket.on('roomsMerged', (destinationTownID: string, requestedTownID: string, destinationFriendlyName: string, 
-    requestedFriendlyName: string , newTownFriendlyName: string, newTownIsPubliclyListed: boolean, newTownIsMergeable: boolean) => {
-    dispatchAppUpdate({ action: 'updateTownToMerge', newTownIDToMerge: destinationTownID});
-
-    let coveyTownID;
-    let friendlyName;
-    if (destinationTownID === video.coveyTownID) {
-      coveyTownID = requestedTownID;
-      friendlyName = requestedFriendlyName;
-    } else {
-      coveyTownID = destinationTownID
-      friendlyName = destinationFriendlyName;
-    }
-    if (toast) {
-      toast({
-        title: 'Town is merging with another town',
-        description: `this town is merging with ${friendlyName} (${coveyTownID}) momentarily! 
-        The new town will be called "${newTownFriendlyName}", 
-        ${newTownIsPubliclyListed ? '✓' : '✗'} Publicly Listed,
-        and ${newTownIsMergeable ? '✓' : '✗'} Mergeable`,
-        status: 'success',
-        isClosable: true,
-        duration: 7000,
-      }) 
-    }
-  })
   const emitMovement = (location: UserLocation) => {
     socket.emit('playerMovement', location);
     dispatchAppUpdate({ action: 'weMoved', location });
@@ -99,7 +195,6 @@ async function GameController(initData: TownJoinResponse,
       myPlayerID: gamePlayerID,
       townIsPubliclyListed: video.isPubliclyListed,
       townIsMergeable: video.isMergeable,
-      townIDToMerge: '',
       emitMovement,
       socket,
       players: initData.currentPlayers.map((sp) => Player.fromServerPlayer(sp)),
@@ -110,12 +205,11 @@ async function GameController(initData: TownJoinResponse,
 
 function App(props: { setOnDisconnect: Dispatch<SetStateAction<Callback | undefined>> }) {
   const [appState, dispatchAppUpdate] = useReducer(appStateReducer, defaultAppState());
-  const toast = useToast();
 
   const setupGameController = useCallback(async (initData: TownJoinResponse) => {
-    await GameController(initData, dispatchAppUpdate, toast);
+    await GameController(initData, dispatchAppUpdate);
     return true;
-  }, [dispatchAppUpdate, toast]);
+  }, [dispatchAppUpdate]);
   const videoInstance = Video.instance();
 
   const { setOnDisconnect } = props;
