+++ frontend/src/components/Login/TownSelection.tsx
@@ -1,4 +1,4 @@
-import React, { useCallback, useEffect, useState, useReducer } from 'react';
+import React, { useCallback, useEffect, useState } from 'react';
 import assert from "assert";
 import {
   Box,
@@ -23,8 +23,8 @@ import useVideoContext from '../VideoCall/VideoFrontend/hooks/useVideoContext/us
 import Video from '../../classes/Video/Video';
 import { CoveyTownInfo, TownJoinResponse, } from '../../classes/TownsServiceClient';
 import useCoveyAppState from '../../hooks/useCoveyAppState';
-import { appStateReducer, defaultAppState } from '../../AppHelper';
 
+// test comment
 interface TownSelectionProps {
   doLogin: (initData: TownJoinResponse) => Promise<boolean>
 }
@@ -37,12 +37,27 @@ export default function TownSelection({ doLogin }: TownSelectionProps): JSX.Elem
   const [townIDToJoin, setTownIDToJoin] = useState<string>('');
   const [currentPublicTowns, setCurrentPublicTowns] = useState<CoveyTownInfo[]>();
   const { connect } = useVideoContext();
-  const { apiClient, townIDToMerge } = useCoveyAppState();
-  // eslint-disable-next-line
-  const [appState, dispatchAppUpdate] = useReducer(appStateReducer, defaultAppState());
+  const { apiClient } = useCoveyAppState();
   const toast = useToast();
 
-  const handleJoin = useCallback(async (coveyRoomID: string, resetToken = false) => {
+  const updateTownListings = useCallback(() => {
+    // console.log(apiClient);
+    apiClient.listTowns()
+      .then((towns) => {
+        setCurrentPublicTowns(towns.towns
+          .sort((a, b) => b.currentOccupancy - a.currentOccupancy)
+        );
+      })
+  }, [setCurrentPublicTowns, apiClient]);
+  useEffect(() => {
+    updateTownListings();
+    const timer = setInterval(updateTownListings, 2000);
+    return () => {
+      clearInterval(timer)
+    };
+  }, [updateTownListings]);
+
+  const handleJoin = useCallback(async (coveyRoomID: string) => {
     try {
       if (!userName || userName.length === 0) {
         toast({
@@ -60,10 +75,7 @@ export default function TownSelection({ doLogin }: TownSelectionProps): JSX.Elem
         });
         return;
       }
-      if (resetToken) {
-        await Video.teardown();
-      }
-      const initData = await Video.setup(userName, coveyRoomID, resetToken);
+      const initData = await Video.setup(userName, coveyRoomID);
 
       const loggedIn = await doLogin(initData);
       if (loggedIn) {
@@ -127,30 +139,6 @@ export default function TownSelection({ doLogin }: TownSelectionProps): JSX.Elem
     }
   };
 
-  const updateTownListings = useCallback(() => {
-    apiClient.listTowns()
-      .then((towns) => {
-        setCurrentPublicTowns(towns.towns
-          .sort((a, b) => b.currentOccupancy - a.currentOccupancy)
-        );
-      })
-  }, [setCurrentPublicTowns, apiClient]);
-  
-  useEffect(() => {
-    updateTownListings();
-    const timer = setInterval(updateTownListings, 2000);
-    return () => {
-      clearInterval(timer)
-    };
-  }, [updateTownListings]);
-
-  useEffect(() => {
-    if (townIDToMerge !== '') {
-      handleJoin(townIDToMerge, true);
-      dispatchAppUpdate({ action: 'updateTownToMerge', newTownIDToMerge: ''});
-    }
-  }, [townIDToMerge, handleJoin]);
-
   return (
     <>
       <form>
@@ -203,13 +191,12 @@ export default function TownSelection({ doLogin }: TownSelectionProps): JSX.Elem
           <Box borderWidth="1px" borderRadius="lg">
             <Heading p="4" as="h2" size="lg">Join an Existing Town</Heading>
             <Box borderWidth="1px" borderRadius="lg">
-              <Flex p="4">
-                <FormControl>
-                  <FormLabel htmlFor="townIDToJoin">Town ID</FormLabel>
-                  <Input name="townIDToJoin" placeholder="ID of town to join, or select from list"
-                        value={townIDToJoin}
-                        onChange={event => setTownIDToJoin(event.target.value)}/>
-                </FormControl>
+              <Flex p="4"><FormControl>
+                <FormLabel htmlFor="townIDToJoin">Town ID</FormLabel>
+                <Input name="townIDToJoin" placeholder="ID of town to join, or select from list"
+                       value={townIDToJoin}
+                       onChange={event => setTownIDToJoin(event.target.value)}/>
+              </FormControl>
                 <Button data-testid='joinTownByIDButton'
                         onClick={() => handleJoin(townIDToJoin)}>Connect</Button>
               </Flex>
