+++ frontend/src/components/Login/TownSelectionPart1.test.tsx
@@ -86,7 +86,6 @@ function wrappedTownSelection() {
     myPlayerID: '',
     currentTownID: '',
     currentTownIsPubliclyListed: false,
-    townIDToMerge: '',
     currentTownIsMergeable: false,
     currentTownFriendlyName: '',
     sessionToken: '',
