+++ frontend/src/components/Login/TownSelectionPart2.test.tsx
@@ -88,7 +88,6 @@ function wrappedTownSelection() {
     currentTownID: '',
     currentTownIsPubliclyListed: false,
     currentTownIsMergeable: false,
-    townIDToMerge: '',
     currentTownFriendlyName: '',
     sessionToken: '',
     userName: '',
@@ -165,7 +164,7 @@ describe('Town Selection - depends on Part 1 passing', () => {
 
         // Check for call sequence
         await waitFor(() => expect(mockVideoSetup)
-          .toBeCalledWith(userName, coveyTownID, false));
+          .toBeCalledWith(userName, coveyTownID));
         await waitFor(() => expect(doLoginMock)
           .toBeCalledWith({ providerVideoToken: videoToken }));
         await waitFor(() => expect(mockConnect)
@@ -220,7 +219,7 @@ describe('Town Selection - depends on Part 1 passing', () => {
 
         // Check for call sequence
         await waitFor(() => expect(mockVideoSetup)
-          .toBeCalledWith(userName, coveyTownID, false));
+          .toBeCalledWith(userName, coveyTownID));
         await waitFor(() => expect(doLoginMock)
           .not
           .toBeCalledWith({ providerVideoToken: videoToken }));
@@ -250,7 +249,7 @@ describe('Town Selection - depends on Part 1 passing', () => {
 
         // Check for call sequence
         await waitFor(() => expect(mockVideoSetup)
-          .toBeCalledWith(userName, coveyTownID, false));
+          .toBeCalledWith(userName, coveyTownID));
         await waitFor(() => expect(doLoginMock)
           .toBeCalledWith({ providerVideoToken: videoToken }));
         await waitFor(() => expect(mockConnect)
@@ -280,7 +279,7 @@ describe('Town Selection - depends on Part 1 passing', () => {
 
         // Check for call sequence
         await waitFor(() => expect(mockVideoSetup)
-          .toBeCalledWith(userName, coveyTownID, false));
+          .toBeCalledWith(userName, coveyTownID));
         await waitFor(() => expect(doLoginMock)
           .toBeCalledWith({ providerVideoToken: videoToken }));
         await waitFor(() => expect(mockConnect)
@@ -319,7 +318,7 @@ describe('Town Selection - depends on Part 1 passing', () => {
               });
               userEvent.click(button);
               await waitFor(() => expect(mockVideoSetup)
-                .toBeCalledWith(username, town.coveyTownID, false));
+                .toBeCalledWith(username, town.coveyTownID));
               await waitFor(() => expect(doLoginMock)
                 .toBeCalledWith({ providerVideoToken: videoToken }));
               await waitFor(() => expect(mockConnect)
@@ -433,7 +432,7 @@ describe('Town Selection - depends on Part 1 passing', () => {
               });
               userEvent.click(button);
               await waitFor(() => expect(mockVideoSetup)
-                .toBeCalledWith(username, town.coveyTownID, false));
+                .toBeCalledWith(username, town.coveyTownID));
               await waitFor(() => expect(doLoginMock)
                 .toBeCalledWith({ providerVideoToken: videoToken }));
               await waitFor(() => expect(mockToast)
@@ -458,7 +457,7 @@ describe('Town Selection - depends on Part 1 passing', () => {
               });
               userEvent.click(button);
               await waitFor(() => expect(mockVideoSetup)
-                .toBeCalledWith(username, town.coveyTownID, false));
+                .toBeCalledWith(username, town.coveyTownID));
               await waitFor(() => expect(doLoginMock)
                 .toBeCalledWith({ providerVideoToken: videoToken }));
               await waitFor(() => expect(mockConnect)
