+++ services/roomService/src/client/CoveyTownsREST.test.ts
@@ -37,7 +37,7 @@ describe('TownsServiceAPIREST', () => {
     const ret = await apiClient.createTown({
       friendlyName,
       isPubliclyListed: isPublic,
-      isMergeable: true,
+      isMergeable: true
     });
     return {
       friendlyName,
