+++ services/roomService/src/client/CoveyTownsSocket.test.ts
@@ -3,6 +3,8 @@ import CORS from 'cors';
 import http from 'http';
 import { nanoid } from 'nanoid';
 import { AddressInfo } from 'net';
+import { promisify } from 'util';
+import io from 'socket.io';
 import * as TestUtils from './TestUtils';
 
 import { UserLocation } from '../CoveyTypes';
@@ -24,7 +26,7 @@ describe('TownServiceApiSocket', () => {
     const ret = await apiClient.createTown({
       friendlyName,
       isPubliclyListed: isPublic,
-      isMergeable: true,
+      isMergeable: true
     });
     return {
       friendlyName,
@@ -34,12 +36,13 @@ describe('TownServiceApiSocket', () => {
     };
   }
 
+  let socketServer: io.Server;
   beforeAll(async () => {
     const app = Express();
     app.use(CORS());
     server = http.createServer(app);
 
-    addTownRoutes(server, app);
+    socketServer = addTownRoutes(server, app);
     server.listen();
     const address = server.address() as AddressInfo;
