+++ services/roomService/src/lib/CoveyTownsStore.test.ts
@@ -20,11 +20,6 @@ function mockCoveyListener(): CoveyTownListener {
     onPlayerJoined(newPlayer: Player) {
       mockCoveyListenerOtherFns(newPlayer);
     },
-    onTownMerged(destinationTownID: string, requestedTownID: string, destinationFriendlyName: string, requestedFriendlyName: string, 
-      newTownFriendlyName: string, newTownIsPubliclyListed: boolean, newTownIsMergeable: boolean){
-      mockCoveyListenerOtherFns(destinationTownID, requestedTownID, destinationFriendlyName, requestedFriendlyName,
-        newTownFriendlyName, newTownIsPubliclyListed, newTownIsMergeable);
-    },
   };
 }
