+++ services/roomService/src/lib/CoveyTownsStore.ts
@@ -27,17 +27,6 @@ export default class CoveyTownsStore {
     return this._towns.find(town => town.coveyTownID === coveyTownID);
   }
 
-  getMergeableTowns(): CoveyTownList {
-    return this._towns.filter(townController => 
-      townController.isPubliclyListed && townController.isMergeable)
-      .map(townController => ({
-        coveyTownID: townController.coveyTownID,
-        friendlyName: townController.friendlyName,
-        currentOccupancy: townController.occupancy,
-        maximumOccupancy: townController.capacity,
-      }));
-  }
-
   getTowns(): CoveyTownList {
     return this._towns.filter(townController => townController.isPubliclyListed)
       .map(townController => ({
@@ -48,13 +37,13 @@ export default class CoveyTownsStore {
       }));
   }
 
-  createTown(friendlyName: string, isPubliclyListed: boolean, isMergeable?: boolean): CoveyTownController {
-    const newTown = new CoveyTownController(friendlyName, isPubliclyListed, isMergeable);
+  createTown(friendlyName: string, isPubliclyListed: boolean): CoveyTownController {
+    const newTown = new CoveyTownController(friendlyName, isPubliclyListed);
     this._towns.push(newTown);
     return newTown;
   }
 
-  updateTown(coveyTownID: string, coveyTownPassword: string, friendlyName?: string, makePublic?: boolean, makeMergeable?: boolean): boolean {
+  updateTown(coveyTownID: string, coveyTownPassword: string, friendlyName?: string, makePublic?: boolean): boolean {
     const existingTown = this.getControllerForTown(coveyTownID);
     if (existingTown && passwordMatches(coveyTownPassword, existingTown.townUpdatePassword)) {
       if (friendlyName !== undefined) {
@@ -66,38 +55,9 @@ export default class CoveyTownsStore {
       if (makePublic !== undefined) {
         existingTown.isPubliclyListed = makePublic;
       }
-      if (makeMergeable !== undefined) {
-        existingTown.isMergeable = makeMergeable;
-      }
       return true;
     }
     return false;
-  }
-
-  mergeTowns(destinationCoveyTownID: string, requestedCoveyTownID: string, 
-    coveyTownPassword: string, newTownFriendlyName: string, 
-    newTownIsPubliclyListed: boolean, newTownIsMergeable: boolean): CoveyTownController | undefined {
-  
-    const destinationTown = this.getControllerForTown(destinationCoveyTownID);
-    const requestedTown = this.getControllerForTown(requestedCoveyTownID);
-    if (destinationTown && passwordMatches(coveyTownPassword, destinationTown.townUpdatePassword)
-      && requestedTown) {
-      destinationTown.townsMerged(destinationTown.coveyTownID, requestedTown.coveyTownID, destinationTown.friendlyName, 
-        requestedTown.friendlyName, newTownFriendlyName, newTownIsPubliclyListed, newTownIsMergeable);
-      requestedTown.townsMerged(destinationTown.coveyTownID, requestedTown.coveyTownID, destinationTown.friendlyName, 
-        requestedTown.friendlyName, newTownFriendlyName, newTownIsPubliclyListed, newTownIsMergeable);
-
-      this.updateTown(destinationTown.coveyTownID, destinationTown.townUpdatePassword, 
-        newTownFriendlyName, newTownIsPubliclyListed, newTownIsMergeable);
-
-      setTimeout(() => {
-        destinationTown.disconnectAllPlayers();
-        this.deleteTown(requestedTown.coveyTownID, requestedTown.townUpdatePassword);
-      }, 7000);
-        
-      return destinationTown;
-    }
-    return undefined;
   }
 
   deleteTown(coveyTownID: string, coveyTownPassword: string): boolean {
