+++ services/roomService/src/server.ts
@@ -16,7 +16,7 @@ server.listen(process.env.PORT || 8081, () => {
   // eslint-disable-next-line no-console
   console.log(`Listening on ${address.port}`);
   if (process.env.DEMO_TOWN_ID) {
-    CoveyTownsStore.getInstance()
+    const newTown = CoveyTownsStore.getInstance()
       .createTown(process.env.DEMO_TOWN_ID, false);
   }
 });
